refactor instrument_deployment routes; now in their own file.  added auth for PUT/POST/DELETE
