+++ ooiservices/app/main/__init__.py
@@ -11,4 +11,4 @@ from flask import Blueprint
 api = Blueprint('main', __name__)
 
 from ooiservices.app.main import routes, authentication, user, operator_event, \
-    annotation
+    annotation, instrument_deployment
